cert-manager: Add cert-manager deployment
Add yaml resources for deploying cert-manager for issuing certificates.

The yaml was taken from
https://github.com/jetstack/cert-manager/releases/download/v1.1.0/cert-manager.yaml
and split up into separate yaml files for easier traceability.

Still to be created is a ClusterIssuer that will get certs from
letsencrypt for the desired domains. A ClusterIssuer is site-specific so
is not added in these generic commits.