metallb: Add MetalLB deployment for LoadBalancers
Add MetalLB (https://metallb.universe.tf) to the cluster to provide
`Services` of type `LoadBalancer`.

The manifests are from:
https://github.com/metallb/metallb/raw/v0.9.5/manifests/namespace.yaml
and
https://github.com/metallb/metallb/raw/v0.9.5/manifests/metallb.yaml

Still to be created is a `ConfigMap` for MetalLB, which contains the IP
address range to use for load balancer addresses. That config is site-
specific so is not added in these generic commits.