sealed-secrets: Add sealed secrets controller
Add the Bitnami Sealed Secrets controller for deploying encrypted
secrets to the cluster.

The manifest is for v0.13.1 from
https://github.com/bitnami-labs/sealed-secrets/releases/download/v0.13.1/controller.yaml

It looks like that yaml manifest is generated from a jsonnet source so
ideally we should look to perhaps deploy that instead of the yaml. But
if we are not customising that, maybe its a little pointless.

Add kubeseal to the tools installed by `make install-tools`.