scripts: Download actual latest version of k3s
Use `sort --version-sort` to figure out what the latest release is
instead of just using the `latest` tag. GitHub tags the latest created
release as the latest, not the highest version. k3s does multiple
simultaneous releases for the different k8s versions so it is a bit
random as to which gets tagged as "latest" by GitHub.