traefik: Add port 8080 as unencrypted http redirect
Add an additional ingress port of 8080 for http traffic that will not
redirect to https. The prior setup had all http traffic (internal and
external) be redirected to https. However some applications running on
k8s require http and not https - in this case it will be the unifi
controller that needs http for the devices to register with the
controller. But other IoT applications may need it due to the low power
IoT devices out there that may not be able to use https.

Expose port 8080 on the internal load balancer service in camh's
configuration. We don't want unencrypted traffic externally, but we can
tolerate it internally.