traefik: Allow TLS to backends with unverified certs
Enable the global `insecureSkipVerify` global setting in traefik that
means connections to backends via TLS do not need to have verifiable
certificates. This is not ideal, but usually we just use http
(unencrypted), so comparatively, this is not too bad.

This is needed for the unifi controller that has its own self-signed
cert and will not accept non-TLS connections. We could generate some
certs, provide them to the unifi controller and to traefik, but this is
easier for now.

To do certs properly, we should probably use cert-manager as a CA,
provide the CA cert to traefik and have cert-manager generate certs on
demand for backends that want them. One day...

As this is a change to the static config, this needs traefik to be
restarted to see the change:

    kubectl delete pod -n traefik -l app=traefik

Docs: https://doc.traefik.io/traefik/routing/overview/#transport-configuration