traefik: Add traefik CRDs
Add the CRDs to the cluster so that traefik configurations can be
created that cannot be done with a standard Ingress - in the coming
case, UDP routers.

Unfortunately there is no clear and simple source for the CRDs (they
dont seem to be bundled with traefik itself (I think the CRD yaml is
generated), but they are unaltered in the helm charts (no helm macros in
these manifests). A make rule has been added to update the CRDs. The
version number of the helm chart has nothing do to with the version
number of traefik unfortunately, so you just need to identify the chart
release for the traefik version and update the CRDs from that tag.

The RBAC privs are even worse. I had to cut and paste them from
https://doc.traefik.io/traefik/reference/dynamic-configuration/kubernetes-crd/
and hope it applies to the version I'm running.

Link: https://doc.traefik.io/traefik/reference/dynamic-configuration/kubernetes-crd/