traefik: Bump to version 2.4.8
Update the traefik deployment to version 2.4.8. This requires a new CRD
for `ServersTransport`, which required bumping the helm version where
the CRDs can be found to v9.17.6.

There were no changes to the `ClusterRole` - that comes from a webpage
and is not versioned, so when I added the CRDs for 2.3, the web page
must have had the `ClusterRole` for 2.4 as it already had
`ServerTransports`.