Add a upgrade target to the Makefile
Add a simple `upgrade` target to the Makefile to upgrade k3s. It just
re-runs the `get-k3s` script (which fetches the latest k3s) and restarts
it. It was enough to upgrade from 1.19 to 1.23 with no issues.

Also update the k3s-killall.sh script as the github shellcheck action is
now complaining about it. Presumably the newer version of shellcheck no
longer likes the script.