trafeik: Update to 2.6.0
Update traefik to 2.6.0 since it seemed to have busted the certs after
being restart after the k3s/k8s upgrade to 1.23 (from 1.19). It looked
like more deprecated api versions that got removed.

This is the usual clusterfuck traefik upgrade - the CRDs now exist in
the helm chart as before, so updating the helm chart version in the
makefile to v10.14.0 was sufficient, except that the RBAC perms needed
to be updated too, and this needs to be downloaded or cut-and-paste
from a web page, and then slightly modified for my deployment since I
have traefik in its own namespace with a simpler name. Argh!!!

I first determined the most recent traefik version to be 2.6.0 by
looking at the github releases page (link 1), and manually update the
`60_deployment.yaml` file. Then I look at the helm chart commit history
(link 2) and look for the version/tag that corresponds to v2.6.0. I
update the Makefile with that version and run

    make update-crds-traefik

Then I go to the YAML web page (link 3) and cut and past the RBAC bit
from the bottom and paste into `12_rbac.yaml` and update the service
account details to

    subjects:
      - kind: ServiceAccount
        name: traefik
        namespace: traefik

so that the name and namespace match my deployment.

Then I can run

    make deploy-traefik

and all is good.

Link: https://github.com/traefik/traefik/releases
Link: https://github.com/traefik/traefik-helm-chart/commits/master
Link: https://doc.traefik.io/traefik/reference/dynamic-configuration/kubernetes-crd/