metallb: Upgrade to 0.10.3 from 0.9.5
Upgrade metallb to 0.10.3 from 0.9.6. Since 0.10.0, metallb now creates
the memberlist secret if it does not exist and is configured to do so.
0.10.X does configure it so, which means we no longer need to manually
create the secret. Remove the memberlist.jsonnet file, bits from the
makefile that deploy it and the stuff in the README about only deploying
the secret once.

Also remove kubecfg from the `install-tools` make target as that was
only needed by the memberlist secret.

The metallb manifest was updated by downloading the yaml from the
repository.

Link: https://github.com/metallb/metallb/blob/v0.10.3/manifests/metallb.yaml
Link: https://metallb.universe.tf/release-notes/#version-0-10-3