metallb: Upgrade to v0.13.3 from v0.12.1
Upgrade metallb to v0.13.3 from v0.12.1. This requires a migration as
metallb now uses custom resources to config it rather than a file in a
configmap. The configs in the `camh` manifest are updated in a
subsequent commit.

The namespace and rest of the k8s resources are now all in one file as
that is how the manifest is now distributed.

Link: https://github.com/metallb/metallb/blob/v0.13.3/config/manifests/metallb-native.yaml
Link: https://metallb.universe.tf/release-notes/#version-0-13-3