sealed-secrets: Reorder manifest resources
Reorder the manifest resources to be in the same order as in the next
release (v0.18.1) so it is easier to see the diff between versions.

I'm not sure if the order is significant or if it may even change
between releases. This reordering was a bit tedious, so we'll see later
if it seemed worth it. It might be better to write a script that splits
out all resources and then imports them in the same order in a top-level
jsonnet file. Soon.