traefik: Upgrade to v2.8.1 from v2.6.1
Upgrade traefik to v2.8.1 just to keep up-to-date.

This is the usual traefik shitfest for upgrading, and I have skipped the
intermediate 2.7.X series as the only point of including the
intermediates is to be able to see what has changed in incremental
steps. But in this case, 2.7.X did a bunch of whitespace changes to the
CRDs which were reverted for 2.8.X, so it's all just noise.

Usual upgrade process:
* Pick version (2.8.1) (link 1)
* Update image tag in `60_deployment.yaml`
* Find the helm chart release corresponding to traefik version (link 2)
* Update Makefile with that version
* Run: `make update-crds-traefik`
* Copy RBAC config from web page (link 3)
* Update the name/namespace for ServiceAccount subject to be
  traefik/traefik to match my deployment
* Deploy with: `make deploy-traefik`

Link: https://github.com/traefik/traefik/releases
Link: https://github.com/traefik/traefik-helm-chart/commits/master
Link: https://doc.traefik.io/traefik/reference/dynamic-configuration/kubernetes-crd/#rbac