metallb: Upgrade to v0.13.7 from v0.13.3
Upgrade metallb to v0.13.7 so we can get the manifest that does not
contain a `PodSecurityPolicy`. This resource was deprecated in
Kubernetes 1.16 (I think) and removed in 1.25. I've just upgraded my k3s
to 1.25 and the metallb manifests no longer apply.

The manifest was updated with:

    curl -o manifests/metallb/manifest.yaml -fsSL \
      'https://raw.githubusercontent.com/metallb/metallb/v0.13.7/config/manifests/metallb-native.yaml'

This release renamed the `ValidatingWebhookConfiguration`
`validating-webhook-configuration` to `metallb-webhook-configuration` so
after applying the new manifest with `make deploy-metallb`, the old
resource needs to be manually deleted:

   kubectl delete validatingwebhookconfiguration validating-webhook-configuration

Link: https://github.com/metallb/metallb/blob/v0.13.7/config/manifests/metallb-native.yaml
Link: https://metallb.universe.tf/release-notes/#version-0-13-7