sealed-secrets: Add update-sealed-secrets recipe
Add a recipe to the Makefile to update the sealed-secrets controller
manifest, and use `kubecfg` to reorder the resources. It seems upstream
generates the `controller.yaml` file in a rather random order that
changes each time a new version comes out, making it hard to identify
the changes. `kubecfg` reorders it in a consistent order making it much
easier to see the changes.