make: Add recipes for updating manifests
Add recipes for fully updating the manifests of `cert-manager`,
`metallb` and `traefik`, similar to what was previously done for
`sealed-secrets`.

`traefik` already had a recipe for updating the CRDs. Extend this to
update the RBAC resources and to update the version in the deployment.

Rename the manifest file for `metallb` to match the config file we
download from upstream so that it is a little easier to reverse engineer
if necessary.

Use a combined manifest for `cert-manager` instead of breaking it up
into small files. Now that we use `kubecfg --order server` to order the
resources consistently, upstream reordering is no longer a problem
solved by small files.. The diff remains relatively small showing just
the changes between versions without whitespace or ordering changes
introduced upstream.

Tweak the previously added `update-sealed-secrets` to match the same
form as these new recipes.