metallb: Switch to auto-generated config
Update the metallb config to an auto-updated version. This has renamed
the config file to be the same as upstream for easier correlation with
upstream.

This is largely a non-functional change. There were some diffs shown
with `kubecfg diff`, and I'm unsure why. I expect I did not properly
deploy last time.

Gen-command: make update-metallb